fix: ch is a char, but sbi_call expects long arguments. On 64-bit systems, passing a 8-bit char where a 64-bit long is expected can cause truncation warnings.. so casting ch to long maybe gonna fix it

+++ RISC-V_OS/kernel.c
@@ -121,7 +121,7 @@ struct sbiret sbi_call(long arg0, long arg1, long arg2, long arg3, long arg4, lo
 }
 
 void putchar(char ch) {
-    sbi_call(ch, 0, 0, 0, 0, 0, 1, 1);
+    sbi_call((long)ch, 0, 0, 0, 0, 0, 1, 1);
 }
 
 extern char __free_ram[], __free_ram_end[];
